[2505-FEAT-371] Add /admin/payments RSC page, migrate types, update nav, tombstone operations
